backend/requirements.txt
Flask==2.3.5
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.6
python-dotenv==1.0.0
